% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR-package.R
\docType{package}
\name{BCsindexRCFS-package}
\alias{BCsindexRCFS}
\alias{BCsindexRCFS-package}
\title{BCsindexRCFS: Site Tools in R}
\description{
A modernized fork of the original SIndexR package that exposes SiteTools' site index routines for practical forestry workflows in R. This package adds consistent wrapper APIs, external DLL backend support, and legacy compatibility guidance. Examples of usage for Permanent Sample Plots (PSP) and growth-and-yield models is provided.
}
\details{
The SiteTools external DLL backend used by this package is published by
  the Government of British Columbia and is available at:
  https://www2.gov.bc.ca/gov/content/industry/forestry/managing-our-forest-resources/forest-inventory/field-forms-and-software/software-download#SiteTools
}
\seealso{
See \code{vignette("workflow-integration", package = "BCsindexRCFS")} for
  practical PSP and treelist workflow examples.

See \code{vignette("legacy-interfaces", package = "BCsindexRCFS")} for a complete
  reference of legacy function mappings and migration guidance.
}
\author{
\strong{Maintainer}: Derek Felix Sattler \email{Derek.Sattler@nrcan-rncan.gc.ca}

Authors:
\itemize{
  \item Derek Felix Sattler \email{Derek.Sattler@nrcan-rncan.gc.ca}
}

Other contributors:
\itemize{
  \item Yong Luo \email{Yong.Luo@gov.bc.ca} (Original SIndexR package author) [contributor]
  \item Ken Polsson (Original Sindex C implementation) [contributor]
}

}
\keyword{internal}
