% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_dll_backend.R
\name{clear_external_dll}
\alias{clear_external_dll}
\alias{SIndexR_ClearExternalDll}
\title{Disable external Sindex DLL backend}
\usage{
clear_external_dll()

SIndexR_ClearExternalDll()
}
\value{
invisible TRUE
}
\description{
Unloads the external DLL and returns wrappers to built-in implementation.
}
\examples{
clear_external_dll()
}
