% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_dll_backend.R
\name{external_dll_info}
\alias{external_dll_info}
\title{External DLL backend status}
\usage{
external_dll_info()
}
\value{
list with `loaded` and `dll_path`
}
\description{
External DLL backend status
}
