% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR_VersionNumber.R
\name{sindex_version}
\alias{sindex_version}
\title{SIndex library version number}
\usage{
sindex_version()
}
\value{
integer version number
}
\description{
Returns the version number of the built-in Sindex routines.
}
\details{
The version is an integer in the form `Mmm` where `M` is the major
release and `mm` is the minor release (e.g. `631` = version 6.31).
If the major release is greater than what your application expects,
assume the Sindex routines are not backward compatible.
}
\examples{
sindex_version()
}
