% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR-package.R
\docType{package}
\name{SIndexR-package}
\alias{SIndexR}
\alias{SIndexR-package}
\title{SIndexR: Sindex in R}
\description{
This package contains key functions in SiteTools.
}
\seealso{
See `vignette("legacy-interfaces")` for a complete reference of legacy
  function mappings and migration guidance.
}
\keyword{internal}
