% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR_CurveList.R
\name{SIndexR_CurveList}
\alias{SIndexR_CurveList}
\title{List site index curves for a species}
\usage{
SIndexR_CurveList(sp_index)
}
\arguments{
\item{sp_index}{Integer/Numeric or character, species index or species code.}
}
\value{
If `sp_index` has length 1, an integer vector of curve indices.
If `sp_index` contains multiple species, a named list of integer vectors.
}
\description{
Returns the defined site index curves for a given species index.
}
\examples{
sp <- SIndexR_FirstSpecies()
SIndexR_CurveList(sp)
SIndexR_CurveList("Sw")
}
