% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_dll_backend.R
\name{SIndexR_ExternalDllInfo}
\alias{SIndexR_ExternalDllInfo}
\title{External DLL backend status}
\usage{
SIndexR_ExternalDllInfo()
}
\value{
list with `loaded` and `dll_path`
}
\description{
External DLL backend status
}
