% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesMapRemap.R
\name{SIndexR_SpecMap}
\alias{SIndexR_SpecMap}
\alias{SIndexR_SpeciesIndex}
\title{Determine species index from species code}
\usage{
SIndexR_SpecMap(sc)

SIndexR_SpeciesIndex(species, fiz = NULL)
}
\arguments{
\item{sc}{Character, Defined as species code.}

\item{species}{Integer or character species index or code.}

\item{fiz}{Character, optional Forest inventory zone code used to disambiguate species codes.}
}
\value{
Species index.
   May return an error code under the following conditions:

   return value    condition
   ------------    ---------
   SI_ERR_CODE     if species code is unknown

Integer species index.
   May return an error code under the following conditions:
   SI_ERR_CODE     if species code is unknown
   SI_ERR_FIZ      if FIZ code is unknown
}
\description{
Determine species index from species code

Convert a species code string or numeric species index into an integer species index.
}
\note{
Species code string can be 1, 2, or 3 letters; upper/lower case
     is ignored.

Species code string can be 1, 2, or 3 letters; upper/lower case is ignored.
}
