% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR-package.R
\docType{package}
\name{SIndexRCFS-package}
\alias{SIndexRCFS}
\alias{SIndexRCFS-package}
\title{SIndexRCFS: Site Index Tools in R for PSP and Growth Workflows}
\description{
A modernized fork of the original SIndexR package that exposes site index routines for practical forestry workflows in R. This package adds consistent wrapper APIs, external DLL backend support, legacy compatibility guidance, and end-to-end examples for Permanent Sample Plot (PSP) and growth-and-yield treelist preparation.
}
\seealso{
See `vignette("legacy-interfaces")` for a complete reference of legacy
  function mappings and migration guidance.
}
\author{
\strong{Maintainer}: Derek Felix Sattler \email{derekfelixsattler@gmail.com}

Authors:
\itemize{
  \item Derek Felix Sattler \email{derekfelixsattler@gmail.com}
}

Other contributors:
\itemize{
  \item Yong Luo (Original SIndexR package author) [contributor]
  \item Ken Polsson (Original Sindex C implementation) [contributor]
}

}
\keyword{internal}
