% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR-package.R
\docType{package}
\name{BCsindexRCFS-package}
\alias{BCsindexRCFS}
\alias{BCsindexRCFS-package}
\title{BCsindexRCFS: Site Tools in R}
\description{
A modernized fork of the original SIndexR package that exposes SiteTools' site index routines for practical forestry workflows in R. This package adds consistent wrapper APIs, external DLL backend support, and legacy compatibility guidance. Examples of usage for Permanent Sample Plots (PSP) and growth-and-yield models is provided.
}
\details{
The SiteTools external DLL backend used by this package is published by
  the Government of British Columbia and is available at:
  https://www2.gov.bc.ca/gov/content/industry/forestry/managing-our-forest-resources/forest-inventory/field-forms-and-software/software-download#SiteTools
}
\section{Definition -- Site Index and Suitable Site Trees}{

  For a given species, site index is defined in BC as the height of the
  largest diameter (at breast height) site tree on a 0.01 ha plot at 50
  years, breast height age, provided the tree meets all the criteria of a
  suitable site tree. A minimum of seven (7) plots, containing suitable site
  trees, in each homogeneous stand stratum is recommended.

  A suitable site tree, for a particular species, reflects the site's full,
  inherent height-growth potential. That is, a suitable site tree is a
  vigorous dominant or co-dominant tree, with a full crown and a straight,
  disease-free, undamaged stem. It cannot be a wolf, open-grown, or veteran
  tree. Furthermore, it must be free from historic influences affecting the
  expression of inherent site potential. This includes both negative effects
  (suppression, repression, damage, disease, etc.) and positive effects
  (fertilization, genetic improvement, etc.).

  Reference: BC Ministry of Forests, Lands, Natural Resource Operations and
  Rural Development (2009). \emph{SIBEC Sampling and Data Standards}.
  \url{http://www2.gov.bc.ca/assets/gov/environment/research-monitoring-and-reporting/research/sibec-documents/standards.pdf}
}

\seealso{
See `vignette("workflow-integration", package = "BCsindexRCFS")` for
  practical PSP and treelist workflow examples.

  See `vignette("legacy-interfaces", package = "BCsindexRCFS")` for a complete
  reference of legacy function mappings and migration guidance.
}
\author{
\strong{Maintainer}: Derek Felix Sattler \email{Derek.Sattler@nrcan-rncan.gc.ca}

Authors:
\itemize{
  \item Derek Felix Sattler \email{Derek.Sattler@nrcan-rncan.gc.ca}
}

Other contributors:
\itemize{
  \item Yong Luo \email{Yong.Luo@gov.bc.ca} (Original SIndexR package author) [contributor]
  \item Ken Polsson (Original Sindex C implementation) [contributor]
}

}
\keyword{package}
