% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_dll_backend.R
\name{clear_external_dll}
\alias{clear_external_dll}
\title{Disable external Sindex DLL backend}
\usage{
clear_external_dll()
}
\value{
invisible TRUE
}
\description{
Unloads the external DLL and returns wrappers to built-in implementation.
}
\examples{
\dontrun{
# After set_external_dll() has been called, revert to built-in:
clear_external_dll()
external_dll_info()$loaded  # FALSE
}
}
