% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_dll_backend.R
\name{sindex_version}
\alias{sindex_version}
\title{SIndex library version number}
\usage{
sindex_version()
}
\value{
integer version number
}
\description{
Returns the version number of the Sindex routines in use. When an external
DLL is loaded via \code{set_external_dll()}, the external DLL version is
returned; otherwise the built-in compiled version is returned.
}
\details{
The version is an integer in the form \code{Mmm} where \code{M} is the major
release and \code{mm} is the minor release (e.g. \code{631} = version 6.31).
}
\examples{
sindex_version()
}
